// packages/shared/src/enums/inventory/stock-location-type.enum.ts
export enum StockLocationType {
  MAIN_WAREHOUSE = 'MAIN_WAREHOUSE',
  MOBILE_TECHNICIAN = 'MOBILE_TECHNICIAN',
  MOBILE_CREW = 'MOBILE_CREW',
  CUSTOMER_SITE = 'CUSTOMER_SITE',
  OFFICE_STOCK = 'OFFICE_STOCK',
  NODE_STOCK = 'NODE_STOCK',
  QUARANTINE = 'QUARANTINE',
  REPAIR = 'REPAIR',
  SCRAP = 'SCRAP',
  INTERNAL_CONSUMPTION = 'INTERNAL_CONSUMPTION',
}